src/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.5
sqlparse==0.5.3
pydantic==2.11.2
